[`FilmEditScreen`] split into separate components
